Short -> int.
Don't arbitrarily limit things.

We can at least stack longs into multiple ints if necessary for storing
things like a block position, or just simply store it as a set of 3
normal integers.

+++ src/main/java/mysticmods/roots/client/KeyHandler.java
@@ -103,9 +103,9 @@ public class KeyHandler {
     }
 
 
-    short max = source.getDataMaximumValue(index);
-    short current = source.getDataValue(spell, index);
-    short newCurrent = current;
+    int max = source.getDataMaximumValue(index);
+    int current = source.getDataValue(spell, index);
+    int newCurrent = current;
 
     if (op == 2 && current < max) {
       newCurrent++;
